instantinsight/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "instantinsight"
version = "0.1.0"
description = "A complete RAG implementation for natural language to SQL conversion using custom agents, AWS Bedrock, and LangChain"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "langchain-core>=0.3.76,<1.0.0",
    "langchain-aws>=0.2.35,<0.3.0",
    "loguru>=0.7.3,<1.0.0",
    "openai>=1.0.0",
    "tiktoken>=0.7.0,<1.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "anthropic>=0.69.0,<1.0.0",
    "python-dotenv",
    "boto3>=1.38.19,<2.0.0",
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn[standard]>=0.23.1,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "networkx>=3.0",
    "langchain-anthropic>=0.3.21,<1.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "pgvector>=0.2.4,<0.3.0",
    "psycopg[binary,pool]>=3.1.0,<4.0.0",
    "PyAthena[SQLAlchemy,pandas]>=3.0.0,<4.0.0",
    "instructor>=1.9.0,<2.0.0",
    "tqdm>=4.66.0,<5.0.0",
    "alembic>=1.16.4,<2.0.0",
    "sqlglot>=27.6.0",
    "fastmcp==2.13.0",
    "httpx>=0.27.0,<1.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "redis>=5.0.1,<6.0.0",
    "hiredis>=2.3.0,<3.0.0",
    "xxhash>=3.4.1,<4.0.0",
    "d3blocks>=1.6.1,<2.0.0",
    "panel>=1.8.1,<2.0.0",
    "strands-agents>=1.10.0",
    "langfuse==2.60.10",
    "openlit>=1.30.0,<2.0.0",
    "ibis-framework[athena,postgres]>=10.6.0",
    "beartype>=0.18.0,<1.0.0",
    "s3fs>=2023.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "pylint>=3.3.7",
    "ipykernel>=6.29.5",
    "vulture>=2.14",
    "ruff>=0.13.0",
]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "requests>=2.32.4",
    "moto[s3]>=5.0.0,<6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]


[tool.pylint.messages_control]
disable = [
    "C",  # Convention - disable all convention checks
    "R",  # Refactoring - disable all refactoring suggestions
    "W0718",  # broad-exception-caught
    "W1203",  # logging-fstring-interpolation
    "E0402",  # import-error
    "W0212",  # protected-access
]

[tool.pylint.format]
max-line-length = 88

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = [
    ".git",
    "__pycache__",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    "build",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D"]
ignore = ["D203", "D212","E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
